Move common drivers to a single place
