+++ src/kernel/src/machine/arm/common/gicv2/mod.rs
@@ -9,7 +9,6 @@
 
 mod gicd;
 mod gicc;
-mod mmio;
 
 use gicd::GICD;
 use gicc::GICC;
