+++ src/kernel/src/machine/arm/common/mod.rs
@@ -1,0 +1,3 @@
+pub mod gicv2;
+pub mod mmio;
+pub mod uart;
