+++ src/kernel/src/machine/arm/virt/serial.rs
@@ -1,7 +1,7 @@
 use lazy_static::lazy_static;
 use twizzler_abi::{device::CacheType, object::Protections};
 
-use super::super::uart::PL011;        
+use super::super::common::uart::PL011;        
 use super::memory::mmio::PL011_UART;
 
 use crate::memory::{VirtAddr, pagetables::{
